Modified Infix_postfix_prefix_operations.java file and renamed Stack.java as StackOperations.java
Modifications in Infix_postfix_prefix_operations.java:
Added a new class called ToInfix which inherits the AdapterClass
In the ToInfix class there will be two methods but the implementation of
only one method is provided.
1) postfixToInfix():
	This method converts the given postfix expression to its
	equivalent infix expression. This method takes one String
	expression as it's parameter. This method is fully implemented.

2) prefixToInfix():
	This method converts the given prefix expression to it's
	equivalent Infix expression. This method also takes one String
	expression as it's parameter. This method will be implemented in
	later commits.

There is also a private method named isOperand() whose return type is
boolean which checks that the symbol of the given expression is an
operand or not.