This code is about the conversion of a number of a base x to the equivalent number of base y. The main method of the code is taking the input as base x,base y and number as a String.
The method Convert is taking the argument the String, base x that is the base of the String and base y is the base of the number to which it is to be converted.
The String num is first converted from the base x to decimal and then the decimal number will be then converted to the equivalent number of the base y.