Added Base_2_to_36.java file and Modified Convertor.java
Base_2_to36.java
----------------------
This Java file has the code to Convert a number of base X to it's
equivalent number of base Y
The method Convert() is bieng overidden here which takes the argument
String number,base x and base y.
Where base x is the base of the number and base y is the base to which
the number is to be converted.
The String number is first Converted from the base X to Decimal and then
the decimal number will be converted to the equivalent number of the
Base Y.

Convertor.java
------------------
In this file a new method Convert() is being introduced in the Convertor
Interface and in the AdapterClass which implements the Interface the
Convert() method is implemented with return value as NULL.